+++ src/gpu/nvidia/cudnn_matmul_impl.hpp
@@ -229,10 +229,10 @@ struct cublas_params : cublas_base_params {
         // bia_mask=1
         if (n_dims == 3) { return convert_dims(dims, new_dims, n_dims); }
         new_dims[0] = 1;
-        for (size_t i = 0; i < n_dims; i++) {
+        for (int i = 0; i < n_dims; i++) {
             new_dims[i + 1] = static_cast<int>(dims[i]);
         }
-        for (size_t i = n_dims; i < 4; i++) {
+        for (int i = n_dims; i < 4; i++) {
             new_dims[i + 1] = 1;
         }
     }
