+++ src/gpu/nvidia/sycl_cuda_utils.hpp
@@ -78,7 +78,7 @@ static void sync_device() {
 
 static void convert_dnnl_dims_array(
         const dnnl_dim_t *dims, int *new_dims, int n_dims) {
-    for (size_t i = 0; i < n_dims; i++) {
+    for (int i = 0; i < n_dims; i++) {
         new_dims[i] = static_cast<int>(dims[i]);
     }
 }
@@ -86,7 +86,7 @@ static void convert_dnnl_dims_array(
 static void convert_dims(const dnnl_dim_t *dims, int *new_dims, int n_dims,
         int adjustment_size = 4, int adjustment_value = 1) {
     convert_dnnl_dims_array(dims, new_dims, n_dims);
-    for (size_t i = n_dims; i < adjustment_size; i++) {
+    for (int i = n_dims; i < adjustment_size; i++) {
         new_dims[i] = adjustment_value;
     }
 }
@@ -135,7 +135,7 @@ static bool adjust_stride_for_dnn(
 
 // Check if the dimensions contain any zeros, returns true if they do.
 static bool has_zero_dims(const dnnl_dim_t *dims, int n_dims) {
-    for (size_t i = 0; i < n_dims; i++) {
+    for (int i = 0; i < n_dims; i++) {
         if (dims[i] == 0) { return true; }
     }
     return false;
